+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Added event handling for new key and existing key
 - Added tests for custom_key_builder and event handling
+- Fix default json coder for datetime.
 
 ### 0.1.6